Add functionality to remove books from the cart
Previously, books could only be added to the cart. This update introduces a "Remove from cart" feature by dispatching actions to update the cart state and subtract the price of the removed item. The button now dynamically handles both adding and removing actions based on the item’s presence in the cart.